requirements.txt
g2pk==0.9.4
torch==1.7.0
numpy==1.19.5
tgt==1.4.3
scipy==1.5.0
pyworld==0.2.10
librosa==0.7.2
numba==0.53.1
matplotlib==3.2.2
unidecode==1.1.1
inflect==4.1.0
g2p-en==2.1.0
tensorboard==2.4.1
python_speech_features==0.6
pandas==1.1.3
pysptk==0.1.18
tensorflow==2.4.0
PyYAML==5.4.1
tqdm==4.46.1
moviepy==1.0.3
quickspacer==1.0.4
